Rotating camera
Script that can be used to set up and alter the camera. Note that the
there is no update in the script, the camera changes when it one of its
parameters is changed. The camera should be able to move to the right,
and the yaw and pitch can be set up.