Rotating Camera Working
I have tested the arcballcamera script on HighRoller and now got it to
work. Only thing to keep in mind is that rotation occurs around the
Y-axis. Adjusting this seems like a lot of work, so we'll have to see
how that pans out. I have included the HighRoller adjusted script as
well so you can see how it works.