Rotating camera FINAL
It finally works! Rigth now you can rotate with WASD keys and the mouse.
To use the mouse you have to press and hold the left mouse button. I
also tried to get it working in the merged Tetris3D, but it doesn't work
as of yet.

The camera class takes two arguments, the vector of the camera position
and the vector of the target. You can set the min and max values of the
Z value of the camera position.